add modules, add webpack config
