+++ index.html
@@ -230,6 +230,6 @@
     </div>
   </template>
 
-  <script src="js/main.js" defer></script>
+  <script src="dist/script.js" defer></script>
 </body>
 </html>
