+++ webpack.config.js
@@ -1,0 +1,9 @@
+const path = require('path');
+
+module.exports = {
+  entry: './js/main.js',
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: 'script.js',
+  },
+};
